requirements.txt
requests
pandas
docopt
tqdm
biopython
bioc
Pillow
tensorflow==2.2.0
keras